Expect string for type attribute in basic update event example test
Previously, the basic update event example expected a numeric value for
the sighting 'type' since the value was specified as a numeric value
when passed into the sighting() client method. With changes in pyMISP
2.4.95, the 'type' value would now be coerced from a numeric value into
a string. This commit adjusts the test to expect the string type value,
for compatibility with the 2.4.95 version.